# JIT: Faster vector == Vector128.Zero on arm64