--- a/src/tests/JIT/HardwareIntrinsics/General/HwiOp/CompareVectorWithZero.cs
+++ b/src/tests/JIT/HardwareIntrinsics/General/HwiOp/CompareVectorWithZero.cs
@@ -1,0 +1,56 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Runtime.CompilerServices;
+using System.Runtime.Intrinsics;
+
+public class CompareVectorWithZero
+{
+    public static int Main()
+    {
+        Test(Vector128.Create(0));
+        Test(Vector128.Create(0.0f));
+        Test(Vector128.Create(-0.0f));
+        Test(Vector128.Create(0.0));
+        Test(Vector128.Create(-0.0));
+        Test(Vector128.Create(-10));
+        Test(Vector128.Create(10));
+        Test(Vector128.Create((sbyte)-10));
+        Test(Vector128.Create((ushort)10));
+        Test(Vector64.Create(0));
+        Test(Vector64.Create(0.0f));
+        Test(Vector64.Create(-0.0f));
+        Test(Vector64.Create(0.0));
+        Test(Vector64.Create(-0.0));
+        Test(Vector64.Create(-10));
+        Test(Vector64.Create(10));
+        Test(Vector64.Create((sbyte)-10));
+        Test(Vector64.Create((ushort)10));
+        Test(Vector128.Create(0, 0, 0, 0, 0, 0, 0, 1));
+        Test(Vector128.Create(0, 0, 0, 0, 0, 0, 0, -1));
+        Test(Vector64.Create(0, 0, 0, 0, 0, 0, 0, 1));
+        Test(Vector64.Create(0, 0, 0, 0, 0, 0, 0, -1));
+        return 100;
+    }
+
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    public static T ToVar<T>(T t) => t;
+
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    public static void AssertTrue(bool expr)
+    {
+        if (!expr)
+            throw new InvalidOperationException();
+    }
+
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    public static void Test<T>(Vector128<T> v) where T : unmanaged =>
+        AssertTrue(v == Vector128<T>.Zero == 
+                   v == Vector128.Create((byte)ToVar(0)));
+
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    public static void Test<T>(Vector64<T> v) where T : unmanaged =>
+        AssertTrue(v == Vector64<T>.Zero == 
+                   v == Vector64.Create((byte)ToVar(0)));
+}