Closes https://github.com/dotnet/runtime/issues/63829
```csharp
static bool IsZero(Vector128<int> vec) => vec == Vector128<int>.Zero;
```
Codegen diff:
```diff
; Assembly listing for method IsZero(System.Runtime.Intrinsics.Vector128`1[Int32]):bool
    stp     fp, lr, [sp,#-16]!
    mov     fp, sp
-   cmeq    v16.4s, v0.4s, #0
-   uminv   b16, v16.16b
-   umov    w0, v16.b[0]
+   umaxv   b16, v0.16b
+   umov    w0, v16.s[0]
    cmp     w0, #0
-   cset    x0, ne
+   cset    x0, eq
    ldp     fp, lr, [sp],#16
    ret     lr
-; Total bytes of code 36
+; Total bytes of code 32
```
This is needed for faster IndexOf from https://github.com/dotnet/runtime/pull/63285. Also, https://github.com/dotnet/runtime/pull/65288 relies on it.

### Perf_Regex_Industry_RustLang_Sherlock [Benchmark](https://gist.github.com/EgorBo/f6edea9a6b1232b05ac1fd5c629b5963):

| Method |               Toolchain |         Pattern |     Mean |    Error |   StdDev | Ratio |
|------- |------------------------ |---------------- |---------:|---------:|---------:|------:|
|  Count |   /Core_Root_PR/corerun | Sherlock Holmes | **68.51 us** | 0.331 us | 0.294 us |  1.00 |
|  Count | /Core_Root_base/corerun | Sherlock Holmes | 72.90 us | 1.352 us | 1.265 us |  1.06 |
|        |                         |                 |          |          |          |       |
|  Count |   /Core_Root_PR/corerun |        sherlock | **56.52 us** | 0.226 us | 0.200 us |  1.00 |
|  Count | /Core_Root_base/corerun |        sherlock | 59.63 us | 0.211 us | 0.198 us |  1.06 |
|        |                         |                 |          |          |          |       |
|  Count |   /Core_Root_PR/corerun |             zqj | **54.03 us** | 0.257 us | 0.241 us |  1.00 |
|  Count | /Core_Root_base/corerun |             zqj | 57.08 us | 0.188 us | 0.176 us |  1.06 |

[Diffs](https://dev.azure.com/dnceng/public/_build/results?buildId=1623470&view=ms.vss-build-web.run-extensions-tab)